requirements.txt
amqp==2.6.1
asgiref==3.3.4
authy==2.2.6
autopep8==1.5.7
billiard==3.6.3.0
celery==4.4.6
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
cryptography==3.4.6
decorator==4.4.2
Django==3.2
django-admin-footer==1.0.1
django-admin-interface==0.14.2
django-admin-logs==1.0.2
django-appconf==1.0.4
django-apptemplates==1.5
django-sage-painless
django-celery-beat==2.2.0
django-colorfield==0.3.2
django-cryptography==1.0
django-cors-headers==3.7.0
django-debug-toolbar==3.2
django-admin-honeypot==1.1.0
django-sage-session==0.1.0
django-dynamic-field==0.0.1
django-defender==0.9.1
django-extensions==3.1.0
django-filter==2.4.0
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-location-field==2.1.0
django-postgres-metrics==0.10.1
django-redis==4.12.1
django-request==1.5.6
django-rest-framework-condition==0.1.1
django-safedelete==0.5.6
django-silk==4.1.0
django-simple-history==3.0.0
django-timezone-field==4.1.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
docutils==0.16
Faker==5.6.0
future==0.18.2
Markdown==3.3.3
MarkupSafe==2.0.1
moviepy==1.0.3
numpy==1.20.3
Pillow==8.1.0
proglog==0.1.9
prompt-toolkit==3.0.16
psycopg2-binary==2.8.6
pycodestyle==2.7.0
pycparser==2.20
pycryptodome==3.10.1
pydotplus==2.0.2
Pygments==2.9.0
PyJWT==2.0.0
pyparsing==2.4.7
python-crontab==2.5.1
python-dateutil==2.8.1
python-decouple==3.4
pytz==2020.5
redis==3.5.3
requests==2.25.1
six==1.15.0
sorl-thumbnail==12.7.0
sqlparse==0.4.1
text-unidecode==1.3
urllib3==1.26.3
vine==1.3.0
django-mapbox-location-field
